Include file 'font5x8.bin'
Apparently, some of the text functions from the adafruit python library throw errors if this is not included. This file does exist in the adafruit library and can clearly be seen on the official github, though for some reason it can not be found during runtime unless it is in the local directory.