Changed the layout of the ON/OFF buttons
After building and running the app on my iPhone, I found that OFF button was running off the screen.... to remedy this, I put the buttons in a Vertical Layout Group. Also, I built the app in a directory called build-ios, so I added that directory name to the .gitignore.